+++ src/utils/index.ts
@@ -2,6 +2,34 @@ import {Calendar} from '../enums/datetime'
 import {Collation} from '../enums/text'
 import {NumberingSystem, Unit} from '../enums/number'
 
+export const getLocalStorageLocale = (): Intl.Locale | null => {
+  const locale = JSON.parse(localStorage.getItem('matriarx_locale') || 'null')
+
+  return locale ? new Intl.Locale(locale) : null
+}
+
+export const setLocalStorageLocale = (locale: Intl.Locale): void =>
+  localStorage.setItem('matriarx_locale', JSON.stringify(locale.baseName))
+
+export const getLocalStorageTranslations = (): Map<
+  string,
+  Map<string, string> | string
+> | null => {
+  const translations = JSON.parse(
+    localStorage.getItem('matriarx_translations') || 'null',
+  )
+
+  return translations
+    ? new Map<string, Map<string, string> | string>(
+        Object.entries(translations),
+      )
+    : null
+}
+
+export const setLocalStorageTranslations = (
+  translations: Map<string, Map<string, string> | string>,
+): void => localStorage.setItem('matriarx_locale', JSON.stringify(translations))
+
 export const getSystemLocale = (): Intl.Locale =>
   new Intl.Locale(
     navigator.languages[0] || 'en-Latn-US',
@@ -29,6 +57,10 @@ export const getSupportedCurrencies = (): string[] =>
   Intl.supportedValuesOf('currency')
 
 export default {
+  getLocalStorageLocale,
+  setLocalStorageLocale,
+  getLocalStorageTranslations,
+  setLocalStorageTranslations,
   getSystemLocale,
   getCanonicalLocales,
   getSupportedCalendars,
